Move scaleModality to HOC
