+++ src/public/common/withScaleAssociation.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import * as d3 from 'd3';
-import { lighten } from 'polished';
+import { mix, complement, lighten } from 'polished';
 import withTheme from '@material-ui/core/styles/withTheme';
 
 function withScaleAssociation(WrappedComponent) {
@@ -16,8 +16,19 @@ function withScaleAssociation(WrappedComponent) {
             theme.palette.primary.main,
           ])
           .unknown('#fff');
+        const tractabilityColor = complement(
+          mix(0.3, theme.palette.primary.main, theme.palette.secondary.main)
+        );
+        const scaleModality = d3
+          .scalePow()
+          .exponent(0.5)
+          .range([lighten(0.4, tractabilityColor), tractabilityColor])
+          .unknown('#fff');
         return (
-          <WrappedComponent scaleAssociation={scaleAssociation} {...rest} />
+          <WrappedComponent
+            {...{ scaleAssociation, scaleModality }}
+            {...rest}
+          />
         );
       }
     }
